Move calculator to flow directory - calculate-your-holiday-entitlement
This calculator is specific to the calculate-your-holiday-entitlement flow and
so should be co-located.

* SmartAnswer::Calculators:HolidayEntitlement
  -> SmartAnswer::CalculateYourHolidayEntitlementFlow::Calculator

* SmartAnswer::Calculators::HolidayEntitlementTest
  -> SmartAnswer::CalculateYourHolidayEntitlementFlow::CalculatorTest

Since these changes do not affect the behaviour and all the regression tests
pass, I've updated the checksums with the new location of the calculator.
