+++ test/data/calculate-your-holiday-entitlement-files.yml
@@ -1,5 +1,5 @@
 ---
-lib/smart_answer_flows/calculate-your-holiday-entitlement.rb: f74f627762a7c41122b950c7b780ab20
+lib/smart_answer_flows/calculate-your-holiday-entitlement.rb: 6d0ebb720e51b05570d9711fde5af72f
 test/data/calculate-your-holiday-entitlement-questions-and-responses.yml: a5d687911e6173e74f2b70af6a5ff7bd
 test/data/calculate-your-holiday-entitlement-responses-and-expected-results.yml: 5ff2290223d8188d45a84cde9883acf6
 lib/smart_answer_flows/calculate-your-holiday-entitlement/calculate_your_holiday_entitlement.govspeak.erb: 76f569916a2429d42e52afc658285f6a
@@ -25,4 +25,4 @@ lib/smart_answer_flows/calculate-your-holiday-entitlement/questions/shift_worker
 lib/smart_answer_flows/calculate-your-holiday-entitlement/questions/what_is_your_leaving_date.govspeak.erb: f5356b35f0d139db060a2092618a99d7
 lib/smart_answer_flows/calculate-your-holiday-entitlement/questions/what_is_your_starting_date.govspeak.erb: d2c6ff5f114bb913455c4db4d4153f9b
 lib/smart_answer_flows/calculate-your-holiday-entitlement/questions/when_does_your_leave_year_start.govspeak.erb: 91de5a5baac16d04d4d0ddc4a5002b96
-lib/smart_answer/calculators/holiday_entitlement.rb: cbcaf72c8cf80eb773d0832c509744e9
+lib/smart_answer_flows/calculate-your-holiday-entitlement/calculator.rb: 7db300412f98e07afa57d92afd453f7d
